Move index fixtures to conftest and add ram_search function
