+++ npcs/memory/search.py
@@ -1,6 +1,7 @@
 import os
 from typing import List
 
+from whoosh.filedb.filestore import RamStorage
 from whoosh.index import FileIndex, create_in, open_dir, exists_in
 from whoosh.qparser import QueryParser
 
@@ -10,9 +11,15 @@ from npcs.utils.constants import ROOT_DIRECTORY
 DEFAULT_SEARCH_RESULT_COUNT = 10
 
 
+# TODO: abstract this interface from the underlying implementation
+# so we can swap Whoosh for FAISS or another index
 def default_index():
     index_path = os.path.join(ROOT_DIRECTORY, 'index')
     return load_index(index_path=index_path, index_name='default')
+
+
+def ram_index():
+    return RamStorage().create_index(NPCMemorySchema)
 
 
 def load_index(index_path: str, index_name: str) -> FileIndex:
@@ -26,7 +33,6 @@ def add_memories(index: FileIndex, data: List[NPCMemory]) -> None:
     # NB: consider using multisegment=True here for initial indexing
     # Not using procs to multithread as it breaks RamStorage (used in tests)
     w = index.writer(limitmb=512)
-    print(data)
     for memory in data:
         w.add_document(**memory.as_document())
     w.commit()
